+++ common/core/src/commonMain/kotlin/database/room/entities/cache/CacheBookByAuthorEntity.kt
@@ -16,7 +16,7 @@ data class CacheBookByAuthorEntity(
     val cacheBook: BookShortRemoteDto?,
     @SerialName("cacheAuthorId") val cacheAuthorId: String,
     @SerialName("cacheTimestamp") val cacheTimestamp: Long,
-    @SerialName("cacheUserId") val cacheUserId: Long,
+    @SerialName("cacheUserId") val cacheUserId: Int,
 )
 
 fun CacheBookByAuthorEntity.toCacheVo(): CacheShortBookByAuthorVo =
