+++ feature/book-info/api/src/commonMain/kotlin/BookInfoRepository.kt
@@ -7,12 +7,12 @@ interface BookInfoRepository {
     suspend fun getLocalBookByLocalId(bookLocalId: Long): Flow<BookVo?>
     suspend fun getLocalBookById(bookId: String): Flow<BookVo?>
     suspend fun updateUserBook(book: BookVo)
-    suspend fun getBookTimestamp(userId: Long): BookTimestampVo
+    suspend fun getBookTimestamp(userId: Int): BookTimestampVo
     suspend fun addOrUpdateLocalBooks(
         books: List<UserBookRemoteDto>,
-        userId: Long
+        userId: Int
     )
 
     suspend fun updateBookTimestamp(lastTimestamp: BookTimestampVo)
-    suspend fun getNotSynchronizedBooks(userId: Long): List<UserBookRemoteDto>
+    suspend fun getNotSynchronizedBooks(userId: Int): List<UserBookRemoteDto>
 }
